requirements.txt
pandas==0.25.2
apache-airflow==1.10.10
boto3===1.12.39